+++ java/src/hype/drawable/HDrawable.java
@@ -1445,6 +1445,22 @@ public abstract class HDrawable extends HNode<HDrawable>
 	
 	// UTILITY //
 	
+	public float x2u(float px) {
+		return px / (_width==0? 100 : _width);
+	}
+	
+	public float y2v(float px) {
+		return px / (_height==0? 100 : _height);
+	}
+	
+	public float u2x(float pc) {
+		return pc * _width;
+	}
+	
+	public float v2y(float pc) {
+		return pc * _height;
+	}
+	
 	/**
 	 * Assigns an _extras bundle_ for this drawable to hold arbitrary data.
 	 * 
@@ -1624,14 +1640,14 @@ public abstract class HDrawable extends HNode<HDrawable>
 	 * This method is primarily used by some subclasses' `draw()` method.
 	 * 
 	 * @param g    The graphics context for this drawable.
-	 * @param currAlphaPerc    The current alpha value in the draw cycle.
+	 * @param alphaPc    The current alpha value in the draw cycle.
 	 */
-	protected void applyStyle(PGraphics g, float currAlphaPerc) {
-		float faPerc = currAlphaPerc * (_fill >>> 24);
+	protected void applyStyle(PGraphics g, float alphaPc) {
+		float faPerc = alphaPc * (_fill >>> 24);
 		g.fill(_fill | 0xFF000000, Math.round(faPerc));
 		
 		if(_strokeWeight > 0) {
-			float saPerc = currAlphaPerc * (_stroke >>> 24);
+			float saPerc = alphaPc * (_stroke >>> 24);
 			g.stroke(_stroke | 0xFF000000, Math.round(saPerc));
 			g.strokeWeight(_strokeWeight);
 			g.strokeCap(_strokeCap);
@@ -1647,9 +1663,9 @@ public abstract class HDrawable extends HNode<HDrawable>
 	 * 
 	 * @param g    The graphics context for this drawable.
 	 * @param usesZ    Indicates if z-coordinates are used.
-	 * @param currAlphaPerc    The current alpha value in the draw cycle.
+	 * @param alphaPc    The current alpha value in the draw cycle.
 	 */
-	public void paintAll(PGraphics g, boolean usesZ, float currAlphaPerc) {
+	public void paintAll(PGraphics g, boolean usesZ, float alphaPc) {
 		if(_alphaPerc<=0) return;
 		g.pushMatrix();
 			// Rotate and translate
@@ -1658,15 +1674,15 @@ public abstract class HDrawable extends HNode<HDrawable>
 			g.rotate(_rotationRad);
 			
 			// Compute current alpha
-			currAlphaPerc *= _alphaPerc;
+			alphaPc *= _alphaPerc;
 			
 			// Draw self
-			draw(g, usesZ,-anchorX(),-anchorY(),currAlphaPerc);
+			draw(g, usesZ,-anchorX(),-anchorY(),alphaPc);
 			
 			// Draw children
 			HDrawable child = _firstChild;
 			while(child != null) {
-				child.paintAll(g, usesZ, currAlphaPerc);
+				child.paintAll(g, usesZ, alphaPc);
 				child = child._next;
 			}
 		g.popMatrix();
@@ -1683,10 +1699,10 @@ public abstract class HDrawable extends HNode<HDrawable>
 	 * @param usesZ Indicates if z-coordinates are used.
 	 * @param drawX The x coordinate where this drawable would consider as 0
 	 * @param drawY The y coordinate where this drawable would consider as 0
-	 * @param currAlphaPerc    The current alpha value in the draw cycle.
+	 * @param alphaPc    The current alpha value in the draw cycle.
 	 */
 	public abstract void draw( PGraphics g, boolean usesZ,
-		float drawX, float drawY, float currAlphaPerc);
+		float drawX, float drawY, float alphaPc);