+++ java/src/hype/drawable/HImage.java
@@ -94,13 +94,13 @@ public class HImage extends HDrawable {
 	
 	@Override
 	public void draw( PGraphics g, boolean usesZ,
-		float drawX, float drawY, float currAlphaPerc
+		float drawX, float drawY, float alphaPc
 	) {
 		if(_image==null) return;
 		
 		// This awkward alpha separation is due to pjs compatibility issues
-		currAlphaPerc *= (_fill>>>24);
-		g.tint( _fill | 0xFF000000, Math.round(currAlphaPerc) );
+		alphaPc *= (_fill>>>24);
+		g.tint( _fill | 0xFF000000, Math.round(alphaPc) );
 		
 		// Determine if the image will be flipped
 		int wscale = 1;