+++ java/src/hype/drawable/HText.java
@@ -144,11 +144,11 @@ public class HText extends HDrawable {
 	
 	@Override
 	public void draw( PGraphics g, boolean usesZ,
-		float drawX, float drawY, float currAlphaPerc
+		float drawX, float drawY, float alphaPc
 	) {
 		if(_text == null) return;
 		
-		applyStyle(g,currAlphaPerc);
+		applyStyle(g,alphaPc);
 		
 		// Determine whether this text will be flipped
 		int wscale = 1;